pyproject.toml
[tool.poetry]
name = "tickets-distributor-bot"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
unidecode = "^1.3.8"
transliterate = "^1.10.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
